Minimap preview on block for field editor
