+++ pxtblocks/fields/field_utils.ts
@@ -68,4 +68,66 @@ namespace pxtblockly {
 
         return canvas.toDataURL();
     }
+
+    export function tilemapToImageURI(data: pxt.sprite.TilemapData, sideLength: number, lightMode: boolean, blocksInfo: pxtc.BlocksInfo) {
+        const colors = pxt.appTarget.runtime.palette.slice();
+        const canvas = document.createElement("canvas");
+        canvas.width = sideLength;
+        canvas.height = sideLength;
+
+        // Works well for all of our default sizes, does not work well if the size is not
+        // a multiple of 2 or is greater than 32 (i.e. from the decompiler)
+        const cellSize = Math.min(sideLength / data.tilemap.width, sideLength / data.tilemap.height);
+
+        // Center the image if it isn't square
+        const xOffset = Math.max(Math.floor((sideLength * (1 - (data.tilemap.width / data.tilemap.height))) / 2), 0);
+        const yOffset = Math.max(Math.floor((sideLength * (1 - (data.tilemap.height / data.tilemap.width))) / 2), 0);
+
+        let context: CanvasRenderingContext2D;
+        if (lightMode) {
+            context = canvas.getContext("2d", { alpha: false });
+            context.fillStyle = "#dedede";
+            context.fillRect(0, 0, sideLength, sideLength);
+        }
+        else {
+            context = canvas.getContext("2d");
+        }
+
+        let tileColors: string[] = [];
+
+        for (let c = 0; c < data.tilemap.width; c++) {
+            for (let r = 0; r < data.tilemap.height; r++) {
+                const tile = data.tilemap.get(c, r);
+
+                if (tile) {
+                    if (!tileColors[tile]) {
+                        const tileInfo = data.tileset.tiles[tile];
+
+                        if (tileInfo.data) {
+                            tileColors[tile] = pxt.sprite.computeAverageColor(pxt.sprite.Bitmap.fromData(tileInfo.data), colors);
+                        }
+                        else {
+                            const bitmap = pxt.sprite.getBitmap(blocksInfo, tileInfo.qualifiedName);
+
+                            if (bitmap) {
+                                tileColors[tile] = pxt.sprite.computeAverageColor(bitmap, colors);
+                            }
+                            else {
+                                tileColors[tile] = "#ffffff";
+                            }
+                        }
+                    }
+
+                    context.fillStyle = tileColors[tile];
+                    context.fillRect(xOffset + c * cellSize, yOffset + r * cellSize, cellSize, cellSize);
+                }
+                else if (lightMode) {
+                    context.fillStyle = "#dedede";
+                    context.fillRect(xOffset + c * cellSize, yOffset + r * cellSize, cellSize, cellSize);
+                }
+            }
+        }
+
+        return canvas.toDataURL();
+    }
 }
