+++ webapp/src/components/ImageEditor/tilemap/Minimap.tsx
@@ -36,6 +36,7 @@ class MinimapImpl extends React.Component<MinimapProps, {}> {
 
         this.canvas.width = bitmap.width;
         this.canvas.height = bitmap.height;
+        this.tileColors = [];
 
         for (let x = 0; x < bitmap.width; x++) {
             for (let y = 0; y < bitmap.height; y++) {
@@ -58,58 +59,13 @@ class MinimapImpl extends React.Component<MinimapProps, {}> {
             if (index >= tileset.tiles.length) {
                 return "#ffffff";
             }
-
-            const averageColor = [0, 0, 0];
-
             const bitmap = pxt.sprite.Bitmap.fromData(tileset.tiles[index].data);
-            const parsedColors = colors.map(colorStringToRGB);
-            let color: number[];
-
-            for (let x = 0; x < bitmap.width; x++) {
-                for (let y = 0; y < bitmap.height; y++) {
-                    color = parsedColors[bitmap.get(x, y)];
-                    averageColor[0] += color[0];
-                    averageColor[1] += color[1];
-                    averageColor[2] += color[2];
-                }
-            }
-
-            const numPixels = bitmap.width * bitmap.height;
-            this.tileColors[index] = "#" + toHex(averageColor.map(c => Math.floor(c / numPixels)));
+            this.tileColors[index] = pxt.sprite.computeAverageColor(bitmap, colors);
         }
 
         return this.tileColors[index];
     }
 }
-
-function colorStringToRGB(color: string) {
-    const parsed = parseColorString(color);
-    return [_r(parsed), _g(parsed), _b(parsed)]
-}
-
-function parseColorString(color: string) {
-    if (color) {
-        if (color.length === 6) {
-            return parseInt("0x" + color);
-        }
-        else if (color.length === 7) {
-            return parseInt("0x" + color.substr(1));
-        }
-    }
-    return 0;
-}
-
-function _r(color: number) { return (color >> 16) & 0xff }
-function _g(color: number) { return (color >> 8) & 0xff }
-function _b(color: number) { return color & 0xff }
-
-function toHex(bytes: ArrayLike<number>) {
-    let r = ""
-    for (let i = 0; i < bytes.length; ++i)
-        r += ("0" + bytes[i].toString(16)).slice(-2)
-    return r
-}
-
 
 
 function mapStateToProps({ store: { present }, editor }: ImageEditorStore, ownProps: any) {
